Fixes and make Async Storage not that slow

+++ chip-testing/src/storage/StorageBackendAsyncJsonFile.ts
@@ -9,6 +9,8 @@ import {
     MaybeAsyncStorage,
     StorageBackendMemory,
     SupportedStorageTypes,
+    Time,
+    createPromise,
     fromJson,
     toJson,
 } from "@matter/general";
@@ -18,6 +20,8 @@ export class StorageBackendAsyncJsonFile extends MaybeAsyncStorage {
     /** We store changes after a value was set to the storage, but not more often than this setting (in ms). */
     private closed = false;
     private store?: StorageBackendMemory;
+    private currentStoreItPromise?: Promise<void>;
+    private lastStoredTime = 0;
 
     constructor(private readonly path: string) {
         super();
@@ -36,6 +40,7 @@ export class StorageBackendAsyncJsonFile extends MaybeAsyncStorage {
         }
         this.store = new StorageBackendMemory(data);
         this.store.initialize();
+        this.lastStoredTime = Time.nowMs();
     }
 
     get initialized() {
@@ -88,15 +93,40 @@ export class StorageBackendAsyncJsonFile extends MaybeAsyncStorage {
             throw new InternalError("Storage not initialized.");
         }
         if (this.closed) return;
+        if (this.currentStoreItPromise !== undefined) {
+            return;
+        }
+        await this.storeIt();
+    }
+
+    private async storeIt(forced = false) {
+        if (this.store === undefined) {
+            throw new InternalError("Storage not initialized.");
+        }
+        if (this.closed) return;
+        if (!forced && this.lastStoredTime < Time.nowMs() - 1000) {
+            return;
+        }
+        if (this.currentStoreItPromise !== undefined) {
+            await this.currentStoreItPromise;
+        }
+
+        const { promise, rejecter, resolver } = createPromise<void>();
         const json = this.toJson(this.store.data);
-        await writeFile(this.path, json, "utf8");
+        writeFile(this.path, json, "utf8")
+            .then(resolver, rejecter)
+            .finally(() => {
+                this.currentStoreItPromise = undefined;
+            });
+        this.currentStoreItPromise = promise;
+        return promise;
     }
 
     override async close() {
         if (this.store === undefined) {
             return;
         }
-        await this.commit();
+        await this.storeIt(true);
         this.closed = true;
         this.store.close();
     }
@@ -127,6 +157,7 @@ export class StorageBackendAsyncJsonFile extends MaybeAsyncStorage {
             throw new InternalError("Storage not initialized.");
         }
         this.store.clearAll(contexts);
+        await this.commit();
     }
 
     private toJson(object: any) {
